Cedar should produce intelligible error messages for type mismatches

+++ Spec/Support/SimpleIncrementer.h
@@ -20,7 +20,7 @@ typedef LargeIncrementerStruct (^ComplexIncrementerBlock)(NSNumber *, LargeIncre
 - (void)incrementBy:(size_t)amount;
 - (void)incrementByNumber:(NSNumber *)number;
 - (void)incrementByInteger:(NSUInteger)number;
-- (void)incrementByABit:(size_t)aBit andABitMore:(NSNumber *)aBitMore;
+- (void)incrementByABit:(unsigned int)aBit andABitMore:(NSNumber *)aBitMore;
 - (void)incrementWithException;
 - (void)methodWithBlock:(void(^)())blockArgument;
 - (void)methodWithCString:(char *)string;
